[refactor] - 리뷰, 메인 query hook정리
