+++ src/components/review/comment/NewComment.tsx
@@ -1,5 +1,6 @@
+import { useNewCommentMutation } from '(@/hooks/useMutation/useCommentMutations)';
 import { useGetUserDataQuery } from '(@/hooks/useQueries/useUserQuery)';
-import { useNewCommentMutation } from '(@/query/review/commentQueryFns)';
+
 import { Button } from '@nextui-org/react';
 import React, { FormEvent, useState } from 'react';
