+++ src/query/review/commentQueryKeys.ts
@@ -1,5 +1,3 @@
 export const COMMENT_AUTHOR_QUERY_KEY = ['comment_data'];
 
 export const COMMENT_QUERY_KEY = ['comment_data'];
-
-export const NEW_COMMENT_QUERY_KEY = ['new_comment'];
